% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaper-package.R
\docType{package}
\name{scaper-package}
\alias{scaper-package}
\alias{_PACKAGE}
\title{scaper: Single Cell Transcriptomics-Level Cytokine Activity Prediction and Estimation}
\description{
Generates cell-level cytokine activity estimates using relevant information from gene sets constructed with the 'CytoSig' and the 'Reactome' databases and scored using the modified 'Variance-adjusted Mahalanobis (VAM)' framework for single-cell RNA-sequencing (scRNA-seq) data. 'CytoSig' database is described in: Jiang at al., (2021) \doi{10.1038/s41592-021-01274-5}. 'Reactome' database is described in: Gillespie et al., (2021) \doi{10.1093/nar/gkab1028}. The 'VAM' method is outlined in: Frost (2020) \doi{10.1093/nar/gkaa582}.
}
\author{
\strong{Maintainer}: Azka Javaid \email{azka.javaid.gr@dartmouth.edu}

Authors:
\itemize{
  \item H. Robert Frost \email{hildreth.r.frost@dartmouth.edu}
}

}
\keyword{internal}
