% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supportedCytokines.R
\name{supportedCytokines}
\alias{supportedCytokines}
\title{Gene set to score for the CytoSig or the Reactome databases.}
\usage{
supportedCytokines(database = "cytosig")
}
\arguments{
\item{database}{Database used for gene set construction and set scoring.
\itemize{
\item "cytosig" returns the 41 cytokines scored using the CytoSig database.
\item "reactome" returns the 30 cytokines scored using the Reactome database.
}}
}
\value{
List of cytokines associated with the CytoSig or the Reactome databases.
}
\description{
Returns the names of the cytokines supported by either the CytoSig or the Reactome databases.
}
\examples{
supportedCytokines(database = "cytosig")
supportedCytokines(database = "reactome")

}
